Mouse events are loaded from input file, Mouse movement events are translated and replayed, with some bugs

+++ src/test/java/BenchmarkingTests.java
@@ -1,4 +1,4 @@
-import edu.temple.UMB.Loader;
+import edu.temple.UMB.KeyLoader;
 import edu.temple.UMB.Recorder;
 import edu.temple.UMB.Replayer;
 import org.junit.jupiter.api.*;
@@ -13,12 +13,14 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.concurrent.*;
 
+
 /**
  * Benchmarks end-to-end record and replay timing characteristics.
  */
 // allows before all annotation to be non-static, see https://docs.junit.org/current/api/org.junit.jupiter.api/org/junit/jupiter/api/TestInstance.Lifecycle.html
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class BenchmarkingTests {
+    boolean benchmarking = false;
     // TODO: test fast keyboard inputs (and mouse!)
     private final String predeterminedEvents = """
 START KEY EVENTS
@@ -51,6 +53,9 @@ EOF
 
     @BeforeAll
     void init() throws RuntimeException, IOException {
+        if (!benchmarking) {
+            return;
+        }
         // before all the tests, setup the file it will read from.
         // check if the tmp dir exists, create if not
         tmpDirFile = new File(tmpDirPath);
@@ -72,6 +77,9 @@ EOF
 
     @AfterAll
     void cleanup() throws IOException {
+        if (!benchmarking) {
+            return;
+        }
         // delete our tmp file and dir
         Files.deleteIfExists(Paths.get(tmpRecorderOutPath));
         Files.deleteIfExists(predeterminedEventsFile.toPath());
@@ -80,6 +88,10 @@ EOF
 
     @Test
     void benchmark() throws InterruptedException {
+        if (!benchmarking) {
+            System.out.println("WARNING: Benchmarking is not enabled!");
+            return;
+        }
         File out = new File(tmpRecorderOutPath);
         // so now we need to set up a replayer, feed it the predetermined events, then set up a recorder.
         // latch allows us to countdown to execution, getting pretty perfect execution times
@@ -117,7 +129,7 @@ EOF
         // there will be some difference in overall timestamps (probably around 50-200ms) due to differences in startup overhead (despite our best efforts to reduce this).
         // that's an important stat, but we really care about variance, or the average gap betweens two events.
         // we can actually just use our loader classes to load the out file
-        Loader l =  new Loader(out);
+        KeyLoader l =  new KeyLoader(out);
         LinkedHashMap<Long, String> recordedEvents;
         try {
             recordedEvents = l.loadJNativeEventsFromFile();
@@ -128,7 +140,7 @@ EOF
 
         // we could just manually parse through the predetermined events string or we can have loader do it for us.
         // need to always remove the last two events of predTS as theyre for the exit escape key and wont get recorded
-        l = new Loader(predeterminedEventsFile);
+        l = new KeyLoader(predeterminedEventsFile);
         LinkedHashMap<Long, String> predEvents;
         try {
             predEvents = l.loadJNativeEventsFromFile();
